service: properly free resources on I/O errors and ctrl+c
In some cases, the master did not tell the services to release their
allocated resources. This would have been done automatically when the
next phase starts, but it now gets done also when there is no next
phase.
This commit also addresses a case where it looked to the master like
a service host got hijacked when ctrl+c was pressed with some service
hosts finished and others still working.