contrib: add mtelbencho (multi test for storage sweep) by Zettar's Chin Fang
Goals and a possible use:
A storage sweep is a simple and effective way to learn about the performance and characteristics of a storage service. Such a sweep should be carried out by any IT professional responsible for an organization's storage, especially for new deployment.

The foremost goal of the mtelbencho wrapper is to simplify the already simple elbencho usage even more for carrying out a storage sweep over a wide range of file sizes. As implemented, it does so by default sweeping from 1KiB to 1TiB, incremented in power-of-two file sizes.  Once installed, please use "mtelbencho.sh -h" on the command line for more details.