contrib/sweep: Introduced a max. open file descriptor check function to the storage sweep tools (#19)
Both mtelbencho.sh and graph_sweep.sh now have a new function set_max_open_file_descriptors(). If the toolset is used in an environment with low ulimit -Hn or ulimit -Sn or both, the low value(s) are automatically set higher so that the toolset can operate error-free. The higher values are used only during the sweep and only for the toolset's use. The system values are left alone.