s3: add options for parallel listing and verification
New option "--s3listobjpar" to list objects in parallel (in contrast to
the previous "--s3listobj"). This parallelizes by using different
listnig prefixes for each thread, based on the datasets created via "-n"
and "-N".
The new option "--s3listverify" verifies such listings of datasets
created via "-n" and "-N" with respect to returned elements and
alphabetic order, so it requires a dataset created with the same values
for "-n" and "-N".