s3: add support for DeleteObjects
Now multiple objects can be deleted in a single request via the new
"--s3multidel" option.
Since this is based on a list+delete loop (instead of relying on known
object names) it can be used to delete arbitrary objects.

This commit also adds the HTTP error code to S3 error messages.