customtree: add option to scan dir/bucket on startup
previously elbencho-scan-path had to be used as a separate tool to
generate a treefile. now there is a new option "--treescan" which can
scan a directory tree or s3 bucket on startup.

this also adds base64 encoding as default to paths in treefiles to
prevent problems with special chars in paths, e.g. line breaks.

this also bumps the version number.