rand: new default rand write algo for max blocks coverage
new default random algo for random write offsets. LCG based to cover
full file range in a single pass, in contrast to "real" random algos
which cover only rather order of 70% of blocks in a single pass.

the previous LCG full coverage random algo had problems with common file
sizes, so this one if a more complex appraoch to eliminate the problems
of the previous approach.

(a different rand offset algo can be selected via "--randalgo".)